adp/dataflow/coderunner/dataflowtools/requirements.txt
tornado==6.4.2
pyparsing==3.1.1
openpyxl==3.0.10
jsonschema==4.20.0
xlrd==2.0.1
xlwt==1.3.0
python-docx==1.1.0
xlutils==2.0.0
requests==2.31.0
httpx==0.26.0
urllib3==2.1.0
python-dateutil==2.8.2
pydantic==2.5.3
pyahocorasick==2.1.0
PyMySQL==1.1.1
environs==11.2.1
pinyin==0.4.0
sqlalchemy==1.3.8
pypandoc==1.15
redis==7.0.1
pypdf==6.1.3